Improve camera property configuration with normalized ranges and proper support detection
- Change CameraConfig brightness/contrast from int (0-100) to double (0-1 normalized)
- Add ConfigureProperty helper method for consistent property handling
- Consolidate ~80 lines of duplicated code into 4 clean ConfigureProperty calls
- First detect property support and ranges, then optionally set config values
- Handle all camera properties (brightness, contrast, saturation, gain) consistently
- Leverage existing DetectPropertyRange caching for optimal performance

+++ src/camera/sources/ICameraSource.h
@@ -75,8 +75,8 @@ struct CameraConfig {
 
     // OpenCV specific settings
     CameraBackend backend = CameraBackend::AUTO;  // Camera backend selection
-    int brightness = -1;                // Camera brightness (-1 = auto)
-    int contrast = -1;                  // Camera contrast (-1 = auto)
+    double brightness = -1.0;             // Camera brightness (-1.0 = auto, 0.0-1.0 normalized)
+    double contrast = -1.0;               // Camera contrast (-1.0 = auto, 0.0-1.0 normalized)
 
     // RealSense specific settings
     bool enableEmitter = true;          // Enable IR emitter for depth
